Add support for custom serialization functions to disk
This generalizes the reading and writing to allows users to use other
methods of serializing objects to disk.

This may be beneficial if the default methods of `readRDS` and `saveRDS`
are too slow, or if the users want to change the amount of compression
used. It also allows the use of faster serialization methods like fst.

Or for specialized cases where you are always caching one type of object
(like data frames) you can use object specific formats like parquet
files.